Clean up Reek exceptions
